Refactor select component
