+++ packages/ui/src/components/account/SelectAccount/OptionListAccount.tsx
@@ -1,9 +1,14 @@
 import React from 'react'
 import { Account } from '../../../common/types'
-import { Option, OptionListProps, OptionsListComponent } from '../../selects'
+import { Option, OptionsListComponent } from '../../selects'
 import { OptionAccount } from './OptionAccount'
 
-export const OptionListAccount = React.memo(({ options, onChange }: OptionListProps<Account>) => (
+interface Props {
+  options: Account[]
+  onChange: (option: Account) => void
+}
+
+export const OptionListAccount = React.memo(({ options, onChange }: Props) => (
   <OptionsListComponent>
     {options.map((option) => (
       <Option key={option.address} onClick={() => onChange && onChange(option)}>
