+++ packages/ui/src/components/selects/components.tsx
@@ -1,9 +1,6 @@
-import React, { ReactNode, useCallback, useEffect, useRef, useState } from 'react'
+import React, { ReactNode } from 'react'
 import styled from 'styled-components'
 import { BorderRad, Colors, Shadows, Sizes, Transitions } from '../../constants'
-import { useToggle } from '../../hooks/useToggle'
-import { Toggle, ToggleButton } from '../buttons/Toggle'
-import { ArrowDownIcon } from '../icons'
 
 interface Props {
   children: ReactNode
@@ -157,100 +154,3 @@ export const OptionsListComponent = styled.ul`
   }
 `
 
-interface SelectProps<T> {
-  disabled?: boolean
-  placeholder: string
-  selected?: T
-  onChange: (selected: T) => void
-  onSearch: (search: string) => void
-  renderSelected: (option: T) => ReactNode
-  renderList: (onOptionClick: (option: T) => void) => ReactNode
-}
-
-export const Select = <T extends any>({
-  disabled,
-  placeholder,
-  selected,
-  onChange,
-  onSearch,
-  renderSelected,
-  renderList,
-}: SelectProps<T>) => {
-  const [filterInput, setFilterInput] = useState('')
-  const search = filterInput
-  const [isOpen, toggleOpen] = useToggle()
-  const [selectedOption, setSelectedOption] = useState<T | undefined>(selected)
-  const selectNode = useRef<HTMLDivElement>(null)
-  const textInput = useRef<HTMLInputElement>(null)
-
-  useEffect(() => {
-    onSearch(search)
-  }, [search])
-
-  const onOptionClick = useCallback(
-    (option: T) => {
-      toggleOpen()
-      setSelectedOption(option)
-      onChange(option)
-      setFilterInput('')
-    },
-    [toggleOpen]
-  )
-
-  useEffect(() => {
-    if (selected) {
-      setSelectedOption(selected)
-      onChange(selected)
-    }
-  }, [selected])
-
-  useEffect(() => {
-    const clickListener = (event: MouseEvent) => {
-      if (isOpen && selectNode.current && !event.composedPath().includes(selectNode.current)) {
-        toggleOpen()
-        setFilterInput('')
-      }
-    }
-    document.addEventListener('mousedown', clickListener)
-
-    return () => document.removeEventListener('mousedown', clickListener)
-  }, [isOpen])
-
-  useEffect(() => {
-    const escListener = (event: KeyboardEvent) => {
-      if (isOpen && event.key === 'Escape') {
-        toggleOpen()
-        setFilterInput('')
-      }
-    }
-    document.addEventListener('keydown', escListener)
-
-    return () => document.removeEventListener('keydown', escListener)
-  }, [isOpen])
-
-  useEffect(() => {
-    isOpen && textInput.current?.focus()
-  }, [isOpen])
-
-  return (
-    <SelectComponent ref={selectNode}>
-      <Toggle onClick={() => !isOpen && toggleOpen()} isOpen={isOpen} disabled={disabled}>
-        {selectedOption && !isOpen && <SelectedOption>{renderSelected(selectedOption)}</SelectedOption>}
-        {(!selectedOption || isOpen) && (
-          <EmptyOption
-            ref={textInput}
-            type="text"
-            placeholder={placeholder}
-            autoComplete="off"
-            value={filterInput}
-            onChange={(t) => setFilterInput(t.target.value)}
-          />
-        )}
-        <ToggleButton disabled={disabled}>
-          <ArrowDownIcon />
-        </ToggleButton>
-      </Toggle>
-      {isOpen && renderList(onOptionClick)}
-    </SelectComponent>
-  )
-}
