+++ packages/ui/src/components/selects/index.ts
@@ -1,2 +1,3 @@
 export * from './components'
 export * from './types'
+export * from './select'
